Fixing loading of categories at the initial page
If the user tried to load categories as the fist page, then the page would load
before the organizations were fetched from the server and no organizations were
displayed. This was because the categories page returned an anonymous function
that captured the value of the orgs as a clojure and did not update when the
original data change. We could fix this two ways, capture the ratom in the
clojure or remove the anonymous function. I choose the
second option because it was the simplest but we could change it to the first if
there is areason to do so.

The bug can be rei-created by hard-reloading the webpage after going to a
catagory.